+++ pages/disclaimer.jsx
@@ -1,3 +1,3 @@
-import { Disclaimer } from 'components/Layout/Disclaimer';
+import { Disclaimer } from 'components/Disclaimer';
 
 export default Disclaimer;
